+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Minor
 
+- Implement `Display` and `Error` for `ValidError` and `EqualError`.
 - Give access to the base of each module.
 
 ### Patch